Assorted changes (2/N)

+++ filters/include/pcl/filters/impl/shadowpoints.hpp
@@ -76,7 +76,7 @@ pcl::ShadowPoints<PointT, NormalT>::applyFilter (PointCloud &output)
   output.points.resize (cp);
   removed_indices_->resize (ri);
   output.width = 1;
-  output.height = static_cast<std::uint32_t> (output.size ());
+  output.height = output.size ();
 }
 
 ///////////////////////////////////////////////////////////////////////////////
